引数の型が間違っていたので修正 空白削除

+++ Source/Actor/Character.h
@@ -81,7 +81,7 @@ public:
     }
     void SetPartsSpecify(
         int index, char* name, float radius,
-        bool cameraTargetFlag, float defensePower, float extractColor)
+        bool cameraTargetFlag, float defensePower, int extractColor)
     {
         parts[index].name = name;
         parts[index].radius = radius;
